+++ FNPlugin/PluginHelper.cs
@@ -146,20 +146,20 @@ namespace FNPlugin
 		private static float _gravityConstant = GameConstants.STANDARD_GRAVITY; 
         public static float GravityConstant { get { return _gravityConstant; } }
 
-		private static double _ispCoreTempMult = GameConstants.IspCoreTemperatureMultiplier;
-        public static double IspCoreTempMult { get { return _ispCoreTempMult; } }
+		private static float _ispCoreTempMult = GameConstants.IspCoreTemperatureMultiplier;
+        public static float IspCoreTempMult { get { return _ispCoreTempMult; } }
 
-		private static double _lowCoreTempBaseThrust = 0;
-        public static double LowCoreTempBaseThrust { get { return _lowCoreTempBaseThrust; } }
+		private static float _lowCoreTempBaseThrust = 0;
+        public static float LowCoreTempBaseThrust { get { return _lowCoreTempBaseThrust; } }
 
-		private static double _highCoreTempThrustMult = GameConstants.HighCoreTempThrustMultiplier;
-        public static double HighCoreTempThrustMult { get { return _highCoreTempThrustMult; } }
+		private static float _highCoreTempThrustMult = GameConstants.HighCoreTempThrustMultiplier;
+        public static float HighCoreTempThrustMult { get { return _highCoreTempThrustMult; } }
 
-		private static double _thrustCoreTempThreshold = 0;
-        public static double ThrustCoreTempThreshold { get { return _thrustCoreTempThreshold; } }
+        private static float _thrustCoreTempThreshold = 0;
+        public static float ThrustCoreTempThreshold { get { return _thrustCoreTempThreshold; } }
 
-		private static double _globalThermalNozzlePowerMaxThrustMult = 1;
-        public static double GlobalThermalNozzlePowerMaxThrustMult { get { return _globalThermalNozzlePowerMaxThrustMult; } }
+		private static float _globalThermalNozzlePowerMaxThrustMult = 1;
+        public static float GlobalThermalNozzlePowerMaxThrustMult { get { return _globalThermalNozzlePowerMaxThrustMult; } }
 
 		private static double _globalMagneticNozzlePowerMaxThrustMult = 1;
         public static double GlobalMagneticNozzlePowerMaxThrustMult { get { return _globalMagneticNozzlePowerMaxThrustMult; } }
@@ -622,7 +622,7 @@ namespace FNPlugin
                     }
                     if (plugin_settings.HasValue("IspCoreTempMult"))
                     {
-                        PluginHelper._ispCoreTempMult = double.Parse(plugin_settings.GetValue("IspCoreTempMult"));
+                        PluginHelper._ispCoreTempMult = float.Parse(plugin_settings.GetValue("IspCoreTempMult"));
                         Debug.Log("[KSP Interstellar] Isp core temperature multiplier set to: " + PluginHelper.IspCoreTempMult.ToString("0.000000"));
                     }
                     if (plugin_settings.HasValue("ElectricEngineIspMult"))
@@ -635,7 +635,7 @@ namespace FNPlugin
 
                     if (plugin_settings.HasValue("GlobalThermalNozzlePowerMaxTrustMult"))
                     {
-                        PluginHelper._globalThermalNozzlePowerMaxThrustMult = double.Parse(plugin_settings.GetValue("GlobalThermalNozzlePowerMaxTrustMult"));
+                        PluginHelper._globalThermalNozzlePowerMaxThrustMult = float.Parse(plugin_settings.GetValue("GlobalThermalNozzlePowerMaxTrustMult"));
                         Debug.Log("[KSP Interstellar] Maximum Global Thermal Power Maximum Thrust Multiplier set to: " + PluginHelper.GlobalThermalNozzlePowerMaxThrustMult.ToString("0.0"));
                     }
                     if (plugin_settings.HasValue("GlobalMagneticNozzlePowerMaxTrustMult"))
@@ -661,17 +661,17 @@ namespace FNPlugin
 
                     if (plugin_settings.HasValue("TrustCoreTempThreshold"))
                     {
-                        PluginHelper._thrustCoreTempThreshold = double.Parse(plugin_settings.GetValue("TrustCoreTempThreshold"));
+                        PluginHelper._thrustCoreTempThreshold = float.Parse(plugin_settings.GetValue("TrustCoreTempThreshold"));
                         Debug.Log("[KSP Interstellar] Thrust core temperature threshold set to: " + PluginHelper.ThrustCoreTempThreshold.ToString("0.0"));
                     }
                     if (plugin_settings.HasValue("LowCoreTempBaseTrust"))
                     {
-                        PluginHelper._lowCoreTempBaseThrust = double.Parse(plugin_settings.GetValue("LowCoreTempBaseTrust"));
+                        PluginHelper._lowCoreTempBaseThrust = float.Parse(plugin_settings.GetValue("LowCoreTempBaseTrust"));
                         Debug.Log("[KSP Interstellar] Low core temperature base thrust modifier set to: " + PluginHelper.LowCoreTempBaseThrust.ToString("0.0"));
                     }
                     if (plugin_settings.HasValue("HighCoreTempTrustMult"))
                     {
-                        PluginHelper._highCoreTempThrustMult = double.Parse(plugin_settings.GetValue("HighCoreTempTrustMult"));
+                        PluginHelper._highCoreTempThrustMult = float.Parse(plugin_settings.GetValue("HighCoreTempTrustMult"));
                         Debug.Log("[KSP Interstellar] High core temperature thrust divider set to: " + PluginHelper.HighCoreTempThrustMult.ToString("0.0"));
                     }
                     if (plugin_settings.HasValue("BasePowerConsumption"))
